+++ db/seeds.rb
@@ -8,14 +8,6 @@
 #   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
-DEVELOPMENT_JUDGE_TEAMS = {
-  "BVAAABSHIRE" => { attorneys: %w[BVAEERDMAN BVARDUBUQUE BVALSHIELDS] },
-  "BVAGSPORER" => { attorneys: %w[BVAOTRANTOW BVAGBOTSFORD BVAJWEHNER1] },
-  "BVAEBECKER" => { attorneys: %w[BVAKBLOCK BVACMERTZ BVAHLUETTGEN] },
-  "BVARERDMAN" => { attorneys: %w[BVASRITCHIE BVAJSCHIMMEL BVAKROHAN1] },
-  "BVAOSCHOWALT" => { attorneys: %w[BVASCASPER1 BVAOWEHNER BVASFUNK1] }
-}.freeze
-
 require "database_cleaner"
 
 # rubocop:disable Metrics/ClassLength
@@ -44,116 +36,7 @@ class SeedDB
     @legacy_appeals.push(LegacyAppeal.create(vacols_id: "reader_id1", vbms_id: "reader_id1"))
   end
 
-  def create_users
-    User.create(css_id: "BVASCASPER1", station_id: 101, full_name: "Steve Attorney_Cases Casper")
-    User.create(css_id: "BVASRITCHIE", station_id: 101, full_name: "Sharree AttorneyNoCases Ritchie")
-    User.create(css_id: "BVAAABSHIRE", station_id: 101, full_name: "Aaron Judge_HearingsAndCases Abshire")
-    User.create(css_id: "BVARERDMAN", station_id: 101, full_name: "Rachael JudgeHasAttorneys_Cases Erdman")
-    User.create(css_id: "BVAEBECKER", station_id: 101, full_name: "Elizabeth Judge_CaseToAssign Becker")
-    User.create(css_id: "BVAKKEELING", station_id: 101, full_name: "Keith Judge_CaseToAssign_NoTeam Keeling")
-    User.create(css_id: "BVATWARNER", station_id: 101, full_name: "Theresa BuildHearingSchedule Warner")
-    User.create(css_id: "BVAGWHITE", station_id: 101, full_name: "George BVADispatchUser_Cases White")
-    User.create(css_id: "BVAGGREY", station_id: 101, full_name: "Gina BVADispatchUser_NoCases Grey")
-    dispatch_admin = User.create(css_id: "BVAGBLACK", station_id: 101, full_name: "Geoffrey BVADispatchAdmin_NoCases Black")
-    OrganizationsUser.make_user_admin(dispatch_admin, BvaDispatch.singleton)
-    case_review_admin = User.create(css_id: "BVAKBLUE", station_id: 101, full_name: "Kim CaseReviewAdmin Blue")
-    OrganizationsUser.make_user_admin(case_review_admin, CaseReview.singleton)
-    special_case_movement_user = User.create(css_id: "BVARDUNKLE",
-                                             station_id: 101,
-                                             full_name: "Rosalie SpecialCaseMovement Dunkle")
-    SpecialCaseMovementTeam.singleton.add_user(special_case_movement_user)
-    special_case_movement_admin = User.create(css_id: "BVAGBEEKMAN",
-                                              station_id: 101,
-                                              full_name: "Bryan SpecialCaseMovementAdmin Beekman")
-    OrganizationsUser.make_user_admin(special_case_movement_admin, SpecialCaseMovementTeam.singleton)
-    bva_intake_admin = User.create(css_id: "BVADWISE", station_id: 101, full_name: "Deborah BvaIntakeAdmin Wise")
-    OrganizationsUser.make_user_admin(bva_intake_admin, BvaIntake.singleton)
-    bva_intake_user = User.create(css_id: "BVAISHAW", station_id: 101, full_name: "Ignacio BvaIntakeUser Shaw")
-    BvaIntake.singleton.add_user(bva_intake_user)
-
-    Functions.grant!("System Admin", users: User.all.pluck(:css_id))
-
-    create_team_admin
-    create_colocated_users
-    create_transcription_team
-    create_vso_users_and_tasks
-    create_field_vso_and_users
-    create_pva_vso_and_users
-    create_org_queue_users
-    create_qr_user
-    create_aod_user_and_tasks
-    create_privacy_user
-    create_lit_support_user
-    create_pulac_cerullo_user
-    create_mail_team_user
-    create_bva_dispatch_user_with_tasks
-    create_case_search_only_user
-    create_judge_teams
-    create_hearings_user_and_tasks
-    create_ama_distribution_tasks
-    create_edit_hearings_user
-    create_non_admin_hearing_coordinator_user
-  end
-
-  def create_ama_distribution_tasks
-    veteran = FactoryBot.create(:veteran, first_name: "Julius", last_name: "Hodge")
-    appeal = FactoryBot.create(:appeal, veteran: veteran, docket_type: Constants.AMA_DOCKETS.evidence_submission)
-    FactoryBot.create(
-      :request_issue,
-      :nonrating,
-      notes: "Pain disorder with 100\% evaluation per examination",
-      decision_review: appeal
-    )
-
-    InitialTasksFactory.new(appeal).create_root_and_sub_tasks!
-
-    # Completing the evidence submission task will mark the appeal ready for distribution
-    evidence_submission_task = EvidenceSubmissionWindowTask.find_by(appeal: appeal)
-    evidence_submission_task.when_timer_ends
-  end
-
-  def create_team_admin
-    u = User.create(css_id: "TEAM_ADMIN", station_id: 101, full_name: "Jim TeamAdminSystemAdmin Jones")
-    existing_sysadmins = Functions.details_for("System Admin")[:granted] || []
-    Functions.grant!("System Admin", users: existing_sysadmins + [u.css_id])
-    Bva.singleton.add_user(u)
-  end
-
-  def create_judge_teams
-    DEVELOPMENT_JUDGE_TEAMS.each_pair do |judge_css_id, h|
-      judge = User.find_or_create_by(css_id: judge_css_id, station_id: 101)
-      judge_team = JudgeTeam.for_judge(judge) || JudgeTeam.create_for_judge(judge)
-      h[:attorneys].each do |css_id|
-        judge_team.add_user(User.find_or_create_by(css_id: css_id, station_id: 101))
-      end
-    end
-  end
-
-  def create_transcription_team
-    transcription_member = User.find_or_create_by(css_id: "TRANSCRIPTION_USER", station_id: 101, full_name: "Noel TranscriptionUser Vasquez")
-    TranscriptionTeam.singleton.add_user(transcription_member)
-  end
-
   ### Hearings Setup ###
-  def create_hearings_user_and_tasks
-    hearings_member = User.find_or_create_by(css_id: "BVATWARNER", station_id: 101)
-    HearingsManagement.singleton.add_user(hearings_member)
-    HearingAdmin.singleton.add_user(hearings_member)
-
-    create_different_hearings_tasks
-    create_change_hearing_disposition_task
-  end
-
-  def create_edit_hearings_user
-    hearings_user = User.create(css_id: "BVASYELLOW", station_id: 101, full_name: "Stacy BuildAndEditHearingSchedule Yellow", roles: ["Edit HearSched", "Build HearSched"])
-    HearingsManagement.singleton.add_user(hearings_user)
-  end
-
-  def create_non_admin_hearing_coordinator_user
-    hearings_user = User.create(css_id: "BVANHALE", station_id: 101, full_name: "Nisha NonAdminHearingCoordinator Hale", roles: ["Edit HearSched"])
-    HearingsManagement.singleton.add_user(hearings_user)
-  end
-
   def create_legacy_case_with_open_schedule_hearing_task(ro_key)
     case ro_key
     when "RO17"
@@ -165,60 +48,6 @@ class SeedDB
     end
 
     LegacyAppeal.find_or_create_by_vacols_id(vacols_id) if vacols_id.present?
-  end
-
-  def create_different_hearings_tasks
-    (%w[RO17 RO19 RO31 RO43 RO45] + [nil]).each do |regional_office|
-      create_legacy_case_with_open_schedule_hearing_task(regional_office)
-
-      rand(10..30).times do
-        appeal = FactoryBot.create(
-          :appeal,
-          :hearing_docket,
-          claimants: [
-            FactoryBot.create(:claimant, participant_id: "CLAIMANT_WITH_PVA_AS_VSO_#{rand(10**10)}")
-          ],
-          closest_regional_office: regional_office
-        )
-
-        FactoryBot.create(:available_hearing_locations, regional_office, appeal: appeal)
-
-        root_task = FactoryBot.create(:root_task, appeal: appeal)
-        distribution_task = FactoryBot.create(
-          :distribution_task,
-          appeal: appeal,
-          parent: root_task
-        )
-        parent_hearing_task = FactoryBot.create(
-          :hearing_task,
-          parent: distribution_task,
-          appeal: appeal
-        )
-
-        schedule_hearing_task_status = [:completed, :in_progress].sample
-
-        FactoryBot.create(
-          :schedule_hearing_task,
-          schedule_hearing_task_status,
-          parent: parent_hearing_task,
-          appeal: appeal
-        )
-
-        # For completed hearing tasks, generate additional tasks too.
-        next unless schedule_hearing_task_status == :completed
-
-        disposition_task = FactoryBot.create(
-          :assign_hearing_disposition_task,
-          parent: parent_hearing_task,
-          appeal: appeal
-        )
-        FactoryBot.create(
-          [:no_show_hearing_task, :evidence_submission_window_task].sample,
-          parent: disposition_task,
-          appeal: appeal
-        )
-      end
-    end
   end
 
   def create_ama_hearing(day)
@@ -331,209 +160,7 @@ class SeedDB
     RequestStore[:current_user] = nil
   end
 
-  def create_change_hearing_disposition_task
-    hearings_member = User.find_or_create_by(css_id: "BVATWARNER", station_id: 101)
-    hearing_day = FactoryBot.create(:hearing_day, created_by: hearings_member, updated_by: hearings_member)
-    veteran = FactoryBot.create(:veteran, first_name: "Abellona", last_name: "Valtas", file_number: 123_456_789)
-    appeal = FactoryBot.create(:appeal, :hearing_docket, veteran_file_number: veteran.file_number)
-    root_task = FactoryBot.create(:root_task, appeal: appeal)
-    distribution_task = FactoryBot.create(:distribution_task, parent: root_task)
-    parent_hearing_task = FactoryBot.create(:hearing_task, parent: distribution_task)
-    FactoryBot.create(:assign_hearing_disposition_task, parent: parent_hearing_task)
-
-    hearing = FactoryBot.create(
-      :hearing,
-      appeal: appeal,
-      hearing_day: hearing_day,
-      created_by: hearings_member,
-      updated_by: hearings_member
-    )
-    FactoryBot.create(:hearing_task_association, hearing: hearing, hearing_task: parent_hearing_task)
-    FactoryBot.create(:change_hearing_disposition_task, parent: parent_hearing_task)
-  end
   ### End Hearings Setup ###
-
-  def create_colocated_users
-    secondary_user = FactoryBot.create(:user, full_name: "Harper SecondaryVLJSupportStaff Tash", roles: %w[Reader])
-    FactoryBot.create(:staff, :colocated_role, user: secondary_user, sdept: "DSP")
-    Colocated.singleton.add_user(secondary_user)
-
-    user = User.create(css_id: "BVALSPORER", station_id: 101, full_name: "Laura Co-located_Cases Sporer", roles: %w[Reader])
-    FactoryBot.create(:staff, :colocated_role, user: user, sdept: "DSP")
-    Colocated.singleton.add_user(user)
-
-    admin = User.create(css_id: "VLJ_SUPPORT_ADMIN", station_id: 101, full_name: "John VLJSupportAdmin Smith", roles: %w[Reader])
-    FactoryBot.create(:staff, :colocated_role, user: admin, sdept: "DSP")
-    OrganizationsUser.make_user_admin(admin, Colocated.singleton)
-  end
-
-  def create_vso_users_and_tasks
-    vso = Vso.create(
-      name: "VSO",
-      url: "veterans-service-organization",
-      participant_id: "2452415"
-    )
-
-    %w[BILLIE MICHAEL].each do |name|
-      u = User.create(
-        css_id: "#{name}_VSO",
-        station_id: 101,
-        full_name: "#{name} VSOUser Jones",
-        roles: %w[VSO]
-      )
-      vso.add_user(u)
-
-      # Assign one IHP task to each member of the VSO team and leave some IHP tasks assigned to the organization.
-      [true, false].each do |assign_to_user|
-        a = FactoryBot.create(:appeal)
-        root_task = FactoryBot.create(:root_task, appeal: a)
-        FactoryBot.create(
-          :hearing,
-          appeal: a
-        )
-        ihp_task = FactoryBot.create(
-          :informal_hearing_presentation_task,
-          parent: root_task,
-          appeal: a,
-          assigned_to: vso
-        )
-        FactoryBot.create(
-          :track_veteran_task,
-          parent: root_task,
-          appeal: a,
-          assigned_to: vso
-        )
-
-        next unless assign_to_user
-
-        InformalHearingPresentationTask.create_many_from_params([{
-                                                                  parent_id: ihp_task.id,
-                                                                  assigned_to_id: u.id,
-                                                                  assigned_to_type: User.name
-                                                                }], u)
-      end
-    end
-  end
-
-  # Creates a VSO org for the PARALYZED VETERANS OF AMERICA VSO that the fake BGS
-  # service returns.
-  #
-  # Use the participant ID `CLAIMANT_WITH_PVA_AS_VSO` to tie this org to a
-  # claimant.
-  def create_pva_vso_and_users
-    vso = Vso.create(
-      name: "PARALYZED VETERANS OF AMERICA, INC.",
-      url: "paralyzed-veteran-of-america",
-      participant_id: "2452383"
-    )
-
-    %w[WINNIE].each do |name|
-      u = User.create(
-        css_id: "#{name}_PVA_VSO",
-        station_id: 101,
-        full_name: "#{name} PVA_VSOUser James",
-        roles: %w[VSO]
-      )
-      vso.add_user(u)
-    end
-  end
-
-  def create_field_vso_and_users
-    vso = FactoryBot.create(:field_vso, name: "Field VSO", url: "field-vso")
-
-    %w[MANDY NICHOLAS ELIJAH].each do |name|
-      u = User.create(
-        css_id: "#{name}_VSO",
-        station_id: 101,
-        full_name: "#{name} VSOUser Wilson",
-        roles: %w[VSO]
-      )
-      vso.add_user(u)
-
-      a = FactoryBot.create(:appeal)
-      root_task = FactoryBot.create(:root_task, appeal: a)
-      FactoryBot.create(
-        :track_veteran_task,
-        parent: root_task,
-        appeal: a,
-        assigned_to: vso
-      )
-    end
-  end
-
-  def create_org_queue_users
-    nca = BusinessLine.create!(name: "National Cemetery Administration", url: "nca")
-    %w[Parveen Chandra Sydney Tai Kennedy].each do |name|
-      u = User.create!(station_id: 101, css_id: "NCA_QUEUE_USER_#{name}", full_name: "#{name} NCAUser Carter")
-      nca.add_user(u)
-    end
-
-    %w[Kun Casey Ariel Naomi Kelly].each do |name|
-      u = User.create!(station_id: 101, css_id: "ORG_QUEUE_USER_#{name}", full_name: "#{name} TranslationUser Cullen")
-      Translation.singleton.add_user(u)
-    end
-  end
-
-  def create_qr_user
-    qr_user = User.create!(station_id: 101, css_id: "QR_USER", full_name: "Yarden QualityReviewer Jordan")
-    QualityReview.singleton.add_user(qr_user)
-
-    # Create QR tasks; one assigned just to the QR org and three assigned both to the org and a QR user.
-    create_task_at_quality_review
-    create_task_at_quality_review(qr_user, "Jane Judge_CaseAtQR Michael", "Joan Attorney_CaseAtQR Ly")
-    create_task_at_quality_review(qr_user, "Cosette Judge_CaseAtQR Zepeda", "Lian Attorney_CaseAtQR Arroyo")
-    create_task_at_quality_review(qr_user, "Huilen Judge_CaseAtQR Concepcion", "Ilva Attorney_CaseAtQR Urrutia")
-  end
-
-  def create_aod_user_and_tasks
-    u = User.create!(station_id: 101, css_id: "AOD_USER", full_name: "Shiloh AODUser Villar")
-    AodTeam.singleton.add_user(u)
-
-    root_task = FactoryBot.create(:root_task)
-    mail_task = ::AodMotionMailTask.create!(
-      appeal: root_task.appeal,
-      parent_id: root_task.id,
-      assigned_to: MailTeam.singleton
-    )
-    ::AodMotionMailTask.create!(
-      appeal: root_task.appeal,
-      parent_id: mail_task.id,
-      assigned_to: AodTeam.singleton
-    )
-  end
-
-  def create_privacy_user
-    u = User.create!(station_id: 101, css_id: "PRIVACY_TEAM_USER", full_name: "Leighton PrivacyAndFOIAUser Naumov")
-    PrivacyTeam.singleton.add_user(u)
-  end
-
-  def create_lit_support_user
-    u = User.create!(station_id: 101, css_id: "LIT_SUPPORT_USER", full_name: "Kiran LitigationSupportUser Rider")
-    LitigationSupport.singleton.add_user(u)
-  end
-
-  def create_pulac_cerullo_user
-    u = User.create!(station_id: 101, css_id: "BVAKSOSNA", full_name: "KATHLEEN PulacCerulloUser SOSNA")
-    PulacCerullo.singleton.add_user(u)
-  end
-
-  def create_mail_team_user
-    u = User.create!(station_id: 101, css_id: "JOLLY_POSTMAN", full_name: "Huan MailUser Tiryaki")
-    MailTeam.singleton.add_user(u)
-  end
-
-  def create_bva_dispatch_user_with_tasks
-    u = User.find_by(css_id: "BVAGWHITE")
-    BvaDispatch.singleton.add_user(u)
-
-    [42, 66, 13].each do |rand_seed|
-      create_task_at_bva_dispatch(rand_seed)
-    end
-  end
-
-  def create_case_search_only_user
-    User.create!(station_id: 101, css_id: "CASE_SEARCHER_ONLY", full_name: "Blair CaseSearchAccessNoQueueAccess Lyon")
-  end
 
   def create_dispatch_tasks(number)
     num_appeals = @legacy_appeals.length
@@ -985,121 +612,6 @@ class SeedDB
       parent: parent,
       appeal: appeal
     )
-  end
-
-  def create_task_at_quality_review(qr_user = nil, judge_name = "Madhu Judge_CaseAtQR Burnham", attorney_name = "Bailey Attorney_CaseAtQR Eoin")
-    vet = FactoryBot.create(
-      :veteran,
-      file_number: Faker::Number.number(digits: 9).to_s,
-      first_name: Faker::Name.first_name,
-      last_name: Faker::Name.last_name
-    )
-    notes = "Pain disorder with 100\% evaluation per examination"
-
-    appeal = FactoryBot.create(
-      :appeal,
-      :with_post_intake_tasks,
-      number_of_claimants: 1,
-      veteran_file_number: vet.file_number,
-      docket_type: Constants.AMA_DOCKETS.direct_review,
-      closest_regional_office: "RO17",
-      request_issues: FactoryBot.create_list(
-        :request_issue, 1, :nonrating, notes: notes
-      )
-    )
-    root_task = appeal.root_task
-
-    judge = FactoryBot.create(:user, station_id: 101)
-    judge.update!(full_name: judge_name) if judge_name
-    FactoryBot.create(:staff, :judge_role, user: judge)
-    judge_task = JudgeAssignTask.create!(appeal: appeal, parent: root_task, assigned_to: judge)
-
-    atty = FactoryBot.create(:user, station_id: 101)
-    atty.update!(full_name: attorney_name) if attorney_name
-    FactoryBot.create(:staff, :attorney_role, user: atty)
-    atty_task_params = { appeal: appeal, parent_id: judge_task.id, assigned_to: atty, assigned_by: judge }
-    atty_task = AttorneyTask.create!(atty_task_params)
-
-    atty_task.update!(status: Constants.TASK_STATUSES.completed)
-    judge_task.update!(status: Constants.TASK_STATUSES.completed)
-
-    qr_org_task = QualityReviewTask.create_from_root_task(root_task)
-
-    if qr_user
-      qr_task_params = [{
-        appeal: appeal,
-        parent_id: qr_org_task.id,
-        assigned_to_id: qr_user.id,
-        assigned_to_type: qr_user.class.name,
-        assigned_by: qr_user
-      }]
-      QualityReviewTask.create_many_from_params(qr_task_params, qr_user).first
-    end
-  end
-
-  def create_task_at_bva_dispatch(seed = Faker::Number.number(digits: 3))
-    Faker::Config.random = Random.new(seed)
-    vet = FactoryBot.create(
-      :veteran,
-      file_number: Faker::Number.number(digits: 9).to_s,
-      first_name: Faker::Name.first_name,
-      last_name: Faker::Name.last_name
-    )
-
-    notes = "Pain disorder with 100\% evaluation per examination"
-
-    appeal = FactoryBot.create(
-      :appeal,
-      :with_post_intake_tasks,
-      number_of_claimants: 1,
-      veteran_file_number: vet.file_number,
-      docket_type: Constants.AMA_DOCKETS.hearing,
-      closest_regional_office: "RO17",
-      request_issues: FactoryBot.create_list(
-        :request_issue, 1, :nonrating, notes: notes
-      )
-    )
-
-    root_task = appeal.root_task
-    judge = FactoryBot.create(:user, station_id: 101, full_name: "Apurva Judge_CaseAtDispatch Wakefield")
-    FactoryBot.create(:staff, :judge_role, user: judge)
-    judge_task = FactoryBot.create(
-      :ama_judge_decision_review_task,
-      assigned_to: judge,
-      appeal: appeal,
-      parent: root_task
-    )
-
-    atty = FactoryBot.create(:user, station_id: 101, full_name: "Andy Attorney_CaseAtDispatch Belanger")
-    FactoryBot.create(:staff, :attorney_role, user: atty)
-    atty_task = FactoryBot.create(
-      :ama_attorney_task,
-      :in_progress,
-      assigned_to: atty,
-      assigned_by: judge,
-      parent: judge_task,
-      appeal: appeal
-    )
-
-    judge_team = JudgeTeam.create_for_judge(judge)
-    judge_team.add_user(atty)
-
-    appeal.request_issues.each do |request_issue|
-      FactoryBot.create(
-        :decision_issue,
-        :nonrating,
-        disposition: "allowed",
-        decision_review: appeal,
-        request_issues: [request_issue],
-        rating_promulgation_date: 2.months.ago,
-        benefit_type: request_issue.benefit_type
-      )
-    end
-
-    atty_task.update!(status: Constants.TASK_STATUSES.completed)
-    judge_task.update!(status: Constants.TASK_STATUSES.completed)
-
-    BvaDispatchTask.create_from_root_task(root_task)
   end
 
   def create_tasks
