+++ db/seeds/base.rb
@@ -1,0 +1,13 @@
+module Seeds
+  class Base
+    private
+
+    def create(*args)
+      FactoryBot.create(*args)
+    end
+
+    def build(*args)
+      FactoryBot.build(*args)
+    end
+  end
+end
