+++ spec/rails_helper.rb
@@ -26,6 +26,9 @@ require "timeout"
 #
 Dir[Rails.root.join("spec/support/**/*.rb")].sort.each { |f| require f }
 
+# because db/seeds is not in the autoload path, we must load them explicitly here
+Dir[Rails.root.join("db/seeds/**/*.rb")].sort.each { |f| require f }
+
 # The TZ variable controls the timezone of the browser in capybara tests, so we always define it.
 # By default (esp for CI) we use Eastern time, so that it doesn't matter where the developer happens to sit.
 ENV["TZ"] ||= "America/New_York"
